Changes to GeneticAlgorithm
Now runs on evaluation set after everything is done. Modified a few
things in the wrapper classes.

+++ KMAD/src/kmad/SpecTuple.java
@@ -9,11 +9,11 @@ package kmad;
  * @author amesen
  */
 public class SpecTuple {
-	int score;
+	double score;
 	int ID;
 	
 	//Constructor
-	public SpecTuple(int pScore, int pID){
+	public SpecTuple(double pScore, int pID){
 		score = pScore;
 		ID = pID;
 	}
